Unify and collectionify read group handling. [WIP]
The reusable macro file has:

 - Defines widgets for common read group parameters.
 - Widgets can be used at top-level (e.g. picard AddOrReplaceReadGroups where handling is required) of inputs or within a conditional (e.g. in the mappers).
 - The conditional (used by mappers) has fours modes:
    - Set according the specification (only ID required - has all parameters). Current BWA approach.
    - Set Picard-style (presumably useful for broad pipelines?) (ID, SM, LB, PU, PL all required - no params for ks, pg, or fo). In terms of what is shown and what is required - mirrors AddOrReplaceReadGroups.
    - Just assign the ID and do so automatically (based either on collection or standalone dataset information). No extra form elements are shown to the user. (Should this be the new default?)
    - Do not set read groups.
 - Utilities for conditionally including parameters and formatting them (eliminates many explicit conditionals in the mappers for instance).
 - An example "mapper" test tool which demonstrates optionally including read groups.
 - An example read group assign test tool which demonstrates unconditionally specifying macros.

In addition to unifying everything, the macro file defines enhanced ID, SM, and LB parameter inputs that are wrapped in a conditional. That conditional allows manual input but also allows the user to just say Galaxy should use collection information or dataset name (if input is not in a collection) to assign these parameters. This should simplify things and make simple workflows more robust, but is absolutely essential for collection-based workflows and newer tool form options that allow running a tool in parallel over many inputs.

This make look at first glance like it is really complicated - but if you focus just on the mappers for instance and treat the macro file like a library - the client interface is actually a good deal easier than doing this stuff by hand for each tool.

I have tests for the macro file, but I haven't tested the tools I update yet. So this is definitely still a work in progress and shouldn't be merged yet, but I did want to pause and give people a chance to complain if they have problems conceptually with the approach.