Unify and collectionify read group handling.
The reusable macro file has:

 - Defines widgets for common read group parameters.
 - Widgets can be used at top-level (e.g. ``picard_AddOrReplaceReadGroups`` where handling is required) of inputs or within a conditional (e.g. in the mappers).
 - The conditional (used by mappers) has fours modes:
    - Set according the specification (only ID required - has all parameters). Current BWA approach.
    - Set Picard-style (presumably useful for broad pipelines?) (``ID``, ``SM``, ``LB``, ``PU``, ``PL`` are all required and there are no params for ``KS``, ``PG``, or ``FO``). In terms of what is shown and what is required - this mirrors ``picard_AddOrReplaceReadGroups``.
    - Just assign the ``ID`` and do so automatically (based either on collection or standalone dataset information). No extra form elements are shown to the user. (Should this be the new default?)
    - Do not set read groups.
 - Utilities for conditionally including parameters and formatting them (eliminates many explicit conditionals in the mappers for instance).
 - An example "mapper" test tool which demonstrates optionally including read groups.
 - An example read group assign test tool which demonstrates unconditionally specifying macros.

In addition to unifying everything, the macro file defines enhanced ``ID``, ``SM``, and ``LB`` parameter inputs that are wrapped in a conditional. That conditional allows manual input but also allows the user to just say Galaxy should use collection information or dataset name (if input is not in a collection) to assign these parameters. This should simplify things and make simple workflows more robust, but is absolutely essential for collection-based workflows and newer tool form options that allow running a tool in parallel over many inputs.

This may look at first glance like it is really complicated - but if you focus just on the mappers for instance and treat the macro file like a library - the client interface is actually a good deal easier than doing this stuff by hand for each tool.

Updates from original version:

 - Tests pass for ``picard_AddOrReplaceReadGroups`` as well as the BWA and Bowtie tools.
 - Remove the changes from tophat - it will require more work (it is less free form and won't accept certain valid read groups arguments).

Smaller Notes:

The new approach for BWA tools excludes empty read groups instead of adding them as empty - hopefully this is fine:

-@RG    ID:rg1  SM:     PL:CAPILLARY
-@PG    ID:bwa  PN:bwa  VN:0.7.12-r1039 CL:bwa sampe -r @RG     ID:rg1  SM:     PL:CAPILLARY localref.fa first.sai second.sai /tmp/tmpn_I5mQ/files/000/dataset_9.dat /tmp/tmpn_I5mQ/files/000/dataset_10.dat
+@RG    ID:rg1  PL:CAPILLARY
+@PG    ID:bwa  PN:bwa  VN:0.7.12-r1039 CL:bwa sampe -r @RG     ID:rg1  PL:CAPILLARY localref.fa first.sai second.sai /tmp/tmpmN3j1j/files/000/dataset_9.dat /tmp/tmpmN3j1j/files/000/dataset_10.dat